Fixed midway tests by injecting constant prior to AngularJS bootstrap
For that, I had to replace the ng-midway-tester library with a modification of my own. It simply introduces a callback function that allows one to execute any Angular dependency injection prior to AngularJS bootstrap call.

There is an AngularJS deferred bootstrap mechanism that we could leverage as well but we'd need to modify the library much further to support it.